Move Websocket context class to its own file as API.js is getting ungainly
